fix: remove useless update
