+++ frontend/src/views/RoomView/TopicsList/ManageTopic/index.tsx
@@ -1,4 +1,4 @@
-import { gql } from "@apollo/client";
+import { gql, useMutation } from "@apollo/client";
 import React, { useCallback } from "react";
 
 import { trackEvent } from "~frontend/analytics/tracking";
@@ -8,11 +8,17 @@ import { useIsCurrentUserTopicManager } from "~frontend/topics/useIsCurrentUserT
 import { CircleOptionsButton } from "~frontend/ui/options/OptionsButton";
 import { openConfirmPrompt } from "~frontend/utils/confirm";
 import { openUIPrompt } from "~frontend/utils/prompt";
-import { useArchiveTopic, useDeleteTopic, useUpdateTopicName } from "~frontend/views/RoomView/TopicsList/shared";
-import { ManageTopic_RoomFragment, ManageTopic_TopicFragment } from "~gql";
+import { useDeleteTopic, useUpdateTopicName } from "~frontend/views/RoomView/TopicsList/shared";
+import {
+  ArchiveTopicMutation,
+  ArchiveTopicMutationVariables,
+  ManageTopic_RoomFragment,
+  ManageTopic_TopicFragment,
+} from "~gql";
 import { createLengthValidator } from "~shared/validation/inputValidation";
 import { IconArchive, IconEdit, IconTrash } from "~ui/icons";
 import { PopoverMenuTrigger } from "~ui/popovers/PopoverMenuTrigger";
+import { addToast } from "~ui/toasts/data";
 
 const fragments = {
   room: gql`
@@ -41,6 +47,27 @@ interface Props {
   topic: ManageTopic_TopicFragment;
   onRenameRequest?: () => void;
 }
+
+export const useArchiveTopic = () =>
+  useMutation<ArchiveTopicMutation, ArchiveTopicMutationVariables & { roomId: string }>(
+    gql`
+      mutation ArchiveTopic($id: uuid!, $archivedAt: timestamptz!) {
+        topic: update_topic_by_pk(pk_columns: { id: $id }, _set: { archived_at: $archivedAt }) {
+          id
+          room_id
+          archived_at
+        }
+      }
+    `,
+    {
+      optimisticResponse: ({ id, roomId, archivedAt }) => ({
+        topic: { __typename: "topic", id, room_id: roomId, archived_at: archivedAt },
+      }),
+      onCompleted() {
+        addToast({ type: "success", title: "Topic was archived" });
+      },
+    }
+  );
 
 export const ManageTopic = withFragments(fragments, ({ room, topic, onRenameRequest }: Props) => {
   const [updateTopicName] = useUpdateTopicName();
